edge_processor/requirements.txt
kafka-python
pandas
MySQL-python
